chore: moving helpers and class renaming
