+++ src/rubrix/server/apis/v0/handlers/text2text.py
@@ -20,7 +20,6 @@ from fastapi import APIRouter, Depends, Query, Security
 from fastapi.responses import StreamingResponse
 
 from rubrix.server.apis.v0.config.tasks_factory import TaskFactory
-from rubrix.server.apis.v0.helpers import takeuntil
 from rubrix.server.apis.v0.models.commons.model import (
     BulkResponse,
     PaginationParams,
@@ -35,6 +34,7 @@ from rubrix.server.apis.v0.models.text2text import (
     Text2TextSearchResults,
 )
 from rubrix.server.errors import EntityNotFoundError
+from rubrix.server.helpers import takeuntil
 from rubrix.server.responses import StreamingResponseWithErrorHandling
 from rubrix.server.security import auth
 from rubrix.server.security.model import User
