+++ src/rubrix/server/apis/v0/handlers/token_classification.py
@@ -21,7 +21,6 @@ from fastapi.responses import StreamingResponse
 
 from rubrix.server.apis.v0.config.tasks_factory import TaskFactory
 from rubrix.server.apis.v0.handlers import token_classification_dataset_settings
-from rubrix.server.apis.v0.helpers import takeuntil
 from rubrix.server.apis.v0.models.commons.model import (
     BulkResponse,
     PaginationParams,
@@ -37,6 +36,7 @@ from rubrix.server.apis.v0.models.token_classification import (
 )
 from rubrix.server.apis.v0.validators.token_classification import DatasetValidator
 from rubrix.server.errors import EntityNotFoundError
+from rubrix.server.helpers import takeuntil
 from rubrix.server.responses import StreamingResponseWithErrorHandling
 from rubrix.server.security import auth
 from rubrix.server.security.model import User
